+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 0.5.0
 
 + `sqlalchemy`: add `unbound_fk()`, `bound_fk()`
-+ Add `timed_cache()`, `TimedDict()`
++ Add `timed_cache()`, `TimedDict()`, `age_and_days()`
++ Add date conversion utilities
 + Move obsolete stuff to `obsolete` package (includes configparse 0.3 as of now)
 + Add more exceptions: `NotFoundError()`